machines: move into seperate component the firmware link details
